worktrack/requirements.txt
asgiref==3.8.1
Django==4.2.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
sqlparse==0.5.2
Markdown==3.7
django-filter==24.3
celery==5.3.1
Faker==33.0.0
python-dotenv==1.0.1
django-celery-beat==2.7.0
django-redis==5.3.0
drf-spectacular==0.27.0
channels==4.2.0
channels_redis==4.2.1 
daphne==4.1.2
